requirements.txt
astroid==2.0.4
atomicwrites==1.2.1
attrs==18.2.0
coverage==4.5.1
docopt==0.6.2
docopt-subcommands==2.3.1
flake8==3.5.0
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
more-itertools==4.3.0
pluggy==0.7.1
py==1.6.0
pychalk==2.0.0
pycodestyle==2.3.1
pyflakes==1.6.0
pylint==2.1.1
pytest==3.8.1
pytest-cov==2.6.0
six==1.11.0
toml==0.9.6
wrapt==1.10.11
